# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HinavAPI",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Default"
    }
  ],
  "paths": {
...
    "/api/authors": {
      "get": {
        "tags": [
          "Author"
        ],
        "summary": "List Authors",
        "description": "著者一覧を取得する\n\n- name: 完全一致検索\n- name_like: 部分一致検索\n- is_favorite: お気に入りフィルタ（True/False/None）",
        "operationId": "list_authors_api_authors_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "nameLike",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namelike"
            }
          },
          {
            "name": "isFavorite",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Isfavorite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorGet"
                  },
                  "title": "Response List Authors Api Authors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Author"
        ],
        "summary": "Update Author",
        "description": "著者情報を更新する（旧形式・非推奨）\n\nこのエンドポイントは後方互換性のために残されています。\n新規実装では PATCH /api/authors/{author_id} を使用してください。",
        "operationId": "update_author_api_authors_patch",
        "deprecated": true,
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAuthor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/authors/{author_id}": {
      "get": {
        "tags": [
          "Author"
        ],
        "summary": "Get Author",
        "description": "著者の詳細情報を取得する\n\n- author_id: 著者ID",
        "operationId": "get_author_api_authors__author_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Author"
        ],
        "summary": "Update Author By Id",
        "description": "著者情報を更新する\n\n- name: 著者名\n- is_favorite: お気に入り設定\n- description: 説明",
        "operationId": "update_author_by_id_api_authors__author_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Author Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Author"
        ],
        "summary": "Delete Author",
        "description": "著者を削除する\n\n書籍に紐づいている著者は削除できません。",
        "operationId": "delete_author_api_authors__author_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_uuid}/authors": {
      "post": {
        "tags": [
          "Author"
        ],
        "summary": "Add Book Author",
        "description": "書籍に著者を追加する\n\n- author_id: 既存の著者IDを指定\n- author_name: 著者名を指定（存在しない場合は新規作成）",
        "operationId": "add_book_author_api_books__book_uuid__authors_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookAuthorPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Author"
        ],
        "summary": "Remove Book Author",
        "description": "書籍から著者を削除する\n\n- author_id: 削除する著者ID",
        "operationId": "remove_book_author_api_books__book_uuid__authors_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookAuthorDelete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "isFavorite": {
            "type": "boolean",
            "title": "Isfavorite"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "bookCount": {
            "type": "integer",
            "title": "Bookcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "isFavorite",
          "description",
          "bookCount"
        ],
        "title": "AuthorDetail",
        "description": "著者詳細レスポンス"
      },
      "AuthorGet": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "isFavorite": {
            "type": "boolean",
            "title": "Isfavorite",
            "default": false
          },
          "bookCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bookcount"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "AuthorGet",
        "description": "著者取得レスポンス"
      },
      "AuthorUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "isFavorite": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isfavorite"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "AuthorUpdate",
        "description": "著者更新リクエスト"
      },
...
      "BookAuthorDelete": {
        "properties": {
          "authorId": {
            "type": "integer",
            "title": "Authorid"
          }
        },
        "type": "object",
        "required": [
          "authorId"
        ],
        "title": "BookAuthorDelete",
        "description": "書籍から著者削除リクエスト"
      },
      "BookAuthorPost": {
        "properties": {
          "authorId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorid"
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorname"
          }
        },
        "type": "object",
        "title": "BookAuthorPost",
        "description": "書籍への著者追加リクエスト"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchAuthor": {
        "properties": {
          "authorId": {
            "type": "integer",
            "title": "Authorid"
          },
          "isFavorite": {
            "type": "boolean",
            "title": "Isfavorite"
          }
        },
        "type": "object",
        "required": [
          "authorId",
          "isFavorite"
        ],
        "title": "PatchAuthor",
        "description": "著者更新リクエスト（旧形式・非推奨）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth"
          }
        }
      }
    }
  },
...
}